Add setting to have units in the header
Add a setting which places all of the units for the measurements in the header, so that the table can be a bit easier to read. Also change the implementation of unit endings to allow for the new setting.

Increase the width of the table to accomodate the heading units.

Add information on the header-units setting to the readme file.

Closes #5